python_handlers/requirements.txt
pydantic>=2.0.0
docker
psycopg2-binary
requests
openstacksdk
keystoneauth1[openidconnect]
python-keystoneclient
